Move speedtest functions to dedicated package
